fix: code restructured, libs does not reference blocks
